requirements.txt
beautifulsoup4
lxml
pandas
requests
scikit-learn
progressbar2
python-dateutil
